feat: added category validations

+++ src/routes/categories.routes.ts
@@ -8,6 +8,12 @@ const categoriesRepository = new CategoriesRepository();
 categoriesRoutes.post("/", (request, response) => {
   const { name, description } = request.body;
 
+  const categoryAlreadyExists = categoriesRepository.findByName(name);
+
+  if (categoryAlreadyExists) {
+    return response.status(400).json({ error: "Category Already Exists" });
+  }
+
   categoriesRepository.create({ name, description });
 
   return response.status(201).send();
@@ -24,11 +30,21 @@ categoriesRoutes.get("/:id", (request, response) => {
 
   const category = categoriesRepository.findById(id);
 
+  if (!category) {
+    return response.status(400).json({ error: "Category Not Found" });
+  }
+
   return response.json(category);
 });
 
 categoriesRoutes.delete("/:id", (request, response) => {
   const { id } = request.params;
+
+  const categoryAlreadyExists = categoriesRepository.findById(id);
+
+  if (!categoryAlreadyExists) {
+    return response.status(400).json({ error: "Category Not Found" });
+  }
 
   categoriesRepository.delete(id);
